feat: v2.3 architecture layer, tests and delivery assets
- new packages: scanner/plugin (C1), eventbus (C2), configvalidation (C3), benchmark (C4), commands (C5), apperror (C6), features, integration tests (C7)
- M1 coverage: api 75.2%, banner 90.8%, network 85.7%; scanner/topology extended tests
- GUI: SRP decomposition (results_view pipeline/render/layout, controllers), theme manager, perf budgets
- network: IPv6 support, hostlist importer, icmp_ping probe
- delivery: Dockerfile, docker-compose, build-docker scripts, systemd/desktop units
- gate: go build ./... clean; go test ./... -short: 48 pkgs ok, 0 FAIL

+++ internal/gui/accent_colors.go
@@ -91,7 +91,10 @@ func (a *App) saveAccentColor(name string, c color.RGBA) {
 		return
 	}
 	p := a.myApp.Preferences()
-	val := c.R<<24 | c.G<<16 | c.B<<8 | c.A
+	// Упаковка RGBA обязана идти в 32 битах: в арифметике uint8 сдвиги R и G
+	// обнулялись, и в Preferences долетали только B и A — акцентный цвет после
+	// перезапуска искажался.
+	val := uint32(c.R)<<24 | uint32(c.G)<<16 | uint32(c.B)<<8 | uint32(c.A)
 	p.SetString(accentKey(name), strconv.FormatUint(uint64(val), 16))
 }
